+++ backend/database/storage/bolt/users.go
@@ -221,9 +221,12 @@ func parseFields(user *users.User, fields []string, actorIsAdmin bool) ([]string
 		for i := 0; i < t.NumField(); i++ {
 			field := t.Field(i)
 			// which=all can't update password
-			if strings.ToLower(field.Name) != "password" && strings.ToLower(field.Name) != "id" && strings.ToLower(field.Name) != "username" {
-				fields = append(fields, field.Name)
+			switch strings.ToLower(field.Name) {
+			case "id", "username", "loginmethod", "password":
+				// Skip these fields
+				continue
 			}
+			fields = append(fields, field.Name)
 		}
 	}
 	newfields := []string{}