+++ backend/swagger/docs/docs.go
@@ -1677,6 +1677,10 @@ const docTemplate = `{
                 "tokenExpirationHours": {
                     "description": "the number of hours until the token expires. Default is 2 hours.",
                     "type": "integer"
+                },
+                "totpSecret": {
+                    "description": "secret used to encrypt TOTP secrets",
+                    "type": "string"
                 }
             }
         },
@@ -1881,6 +1885,7 @@ const docTemplate = `{
                     "type": "boolean"
                 },
                 "minLength": {
+                    "description": "minimum pasword length required.",
                     "type": "integer",
                     "minimum": 5
                 },
@@ -1893,7 +1898,11 @@ const docTemplate = `{
                     ]
                 },
                 "signup": {
-                    "description": "currently not used by filebrowser",
+                    "description": "allow signups on login page if enabled -- not secure.",
+                    "type": "boolean"
+                },
+                "totpEnforced": {
+                    "description": "if set to true, TOTP is enforced for all password users users. Otherwise, users can choose to enable TOTP.",
                     "type": "boolean"
                 }
             }
@@ -2368,6 +2377,10 @@ const docTemplate = `{
                 "loginMethod": {
                     "$ref": "#/definitions/users.LoginMethod"
                 },
+                "otpEnabled": {
+                    "description": "true if TOTP is enabled, false otherwise",
+                    "type": "boolean"
+                },
                 "password": {
                     "type": "string"
                 },
@@ -2415,6 +2428,12 @@ const docTemplate = `{
                 },
                 "themeColor": {
                     "description": "theme color to use: eg. #ff0000, or var(--red), var(--purple), etc",
+                    "type": "string"
+                },
+                "totpNonce": {
+                    "type": "string"
+                },
+                "totpSecret": {
                     "type": "string"
                 },
                 "username": {