+++ frontend/src/views/Login.vue
@@ -8,7 +8,7 @@
         <h3>{{ loginName }}</h3>
       </div>
       <div v-if="passwordAvailable" class="password-entry">
-        <div v-if="error !== ''" class="wrong">{{ error }}</div>
+        <div v-if="error !== ''" class="wrong-login">{{ error }}</div>
         <input
           autofocus
           class="input input--block"
@@ -50,13 +50,16 @@
       </div>
     </form>
   </div>
+  <prompts :class="{ 'dark-mode': isDarkMode }"></prompts>
 </template>
 
 <script>
 import router from "@/router";
-import { state } from "@/store";
+import { mutations, state } from "@/store";
+import Prompts from "@/components/prompts/Prompts.vue";
 import Icon from "@/components/files/Icon.vue";
-import { signupLogin, login, initAuth } from "@/utils/auth";
+import { usersApi } from "@/api";
+import { initAuth } from "@/utils/auth";
 import {
   name,
   logoURL,
@@ -72,6 +75,7 @@ export default {
   name: "login",
   components: {
     Icon,
+    Prompts,
   },
   computed: {
     signup: () => signup,
@@ -133,16 +137,24 @@ export default {
       }
       try {
         if (this.createMode) {
-          await signupLogin(this.username, this.password);
+          await usersApi.signupLogin(this.username, this.password);
         }
-        console.log("Logging in...");
-        await login(this.username, this.password, captcha);
-        console.log("Logged in");
+        await usersApi.login(this.username, this.password, captcha);
         await initAuth();
-
         router.push({ path: redirect });
       } catch (e) {
-        console.error(e);
+        if (e.message == 403) {
+          console.log("showing TOTP prompt");
+          mutations.showHover({
+            name: "totp",
+            props: {
+              username: this.username,
+              password: this.password,
+              recaptcha: captcha,
+              redirect: redirect,
+            },
+          });
+        }
         if (e.message == 409) {
           this.error = this.$t("login.usernameTaken");
         } else {